+++ src/test/scala/org/broadinstitute/gpp/poolq3/parser/ReferenceEntryTest.scala
@@ -9,12 +9,20 @@ import munit.FunSuite
 
 class ReferenceEntryTest extends FunSuite:
 
-  test("barcodeLengths detects splits") {
+  test("barcodeLengths detects unambiguous splits") {
     val cs = "C" * 10
     val ts = "T" * 8
     val bc = s"$cs:$ts"
     val re = ReferenceEntry(bc, "Some Cs and some Ts")
-    assertEquals(re.barcodeLengths, (10, 8))
+    assertEquals(re.barcodeLengths, Some((10, 8)))
+  }
+
+  test("barcodeLengths rejects ambiguous splits") {
+    val cs = "C" * 10
+    val ts = "T" * 8
+    val bc = s"$cs:$ts;"
+    val re = ReferenceEntry(bc, "Some Cs and some Ts")
+    assertEquals(re.barcodeLengths, None)
   }
 
 end ReferenceEntryTest